Refractored reset using Verification class
